requirements.txt
PyMuPDF
tqdm
pandas
joblib
torch
transformers
spacypdfreader
spacy
pdfplumber
sentence-transformers
numpy
google-generativeai
python-dotenv
requests